% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-data.R
\name{delphi_palettes}
\alias{delphi_palettes}
\title{Delphi palette collection}
\usage{
delphi_palettes(category = NULL, colorblind_friendly = FALSE)
}
\arguments{
\item{category}{An optional category. It must be one of \code{classic}, \code{nature},
\code{keycaps}, \code{vintage}, \code{modern}, \code{bold}, \code{soft}, \code{monochrome}, \code{seasonal},
or \code{artistic}.}

\item{colorblind_friendly}{If \code{TRUE}, returns only palettes that pass
\code{\link[=delphi_palette_colorblind]{delphi_palette_colorblind()}}.}
}
\value{
A named list. Each element is a character vector of hexadecimal
colors.
}
\description{
Returns the Delphi palette collection. Palette names use lowercase letters
and hyphens. For example, use \code{mayfair} or \code{perito-moreno}.
}
\examples{
delphi_palettes("nature")
delphi_palettes(colorblind_friendly = TRUE)
}
