% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{delphi_palette}
\alias{delphi_palette}
\title{Get a Delphi palette}
\usage{
delphi_palette(name = "mayfair", n = NULL, type = "discrete", direction = 1)
}
\arguments{
\item{name}{A single palette name. Use \code{\link[=delphi_palettes]{delphi_palettes()}} to get palette
names. The default is \code{"mayfair"}.}

\item{n}{The number of colors to return. The default returns all colors.
You must set \code{n} for \code{type = "continuous"}.}

\item{type}{The palette type. Use \code{"discrete"} for stored colors. Use
\code{"continuous"} for interpolated colors.}

\item{direction}{The color order. Use \code{1} for the stored order. Use \code{-1} to
reverse the stored order before this function selects \code{n} colors.}
}
\value{
A character vector of hexadecimal colors.
}
\description{
Returns colors from one Delphi palette. The default palette is \code{mayfair}.
Set \code{type = "continuous"} and \code{n} to interpolate colors.
}
\examples{
delphi_palette("mayfair")
delphi_palette("mayfair", n = 8, type = "continuous")
delphi_palette(direction = -1)
}
