% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphiPalettes-package.R
\docType{package}
\name{delphiPalettes-package}
\alias{delphiPalettes}
\alias{delphiPalettes-package}
\title{delphiPalettes: Color Palettes Curated by 'delphi.tools'}
\description{
Provides color palettes from the collection at 'delphi.tools' \url{https://delphi.tools}. delphi creates privacy-respecting design tools \url{https://github.com/1612elphi/delphitools}. The package groups palettes by category. It returns palette colors or interpolated colors. It also reports whether a palette keeps distinct colors under simulated color vision differences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ConnorB/delphiPalettes}
  \item \url{https://connorb.github.io/delphiPalettes/}
  \item Report bugs at \url{https://github.com/ConnorB/delphiPalettes/issues}
}

}
\author{
\strong{Maintainer}: Connor Brown \email{ConnorBrown1996@gmail.com} (\href{https://orcid.org/0000-0002-9680-8930}{ORCID})

Authors:
\itemize{
  \item Connor Brown \email{ConnorBrown1996@gmail.com} (\href{https://orcid.org/0000-0002-9680-8930}{ORCID})
}

Other contributors:
\itemize{
  \item Ruby Morgan Voigt (Original palette designs ported from https://delphi.tools (GitHub: 1612elphi)) [data contributor]
}

}
\keyword{internal}
