% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_delphi}
\alias{scale_delphi}
\alias{scale_color_delphi_d}
\alias{scale_fill_delphi_d}
\alias{scale_color_delphi_c}
\alias{scale_fill_delphi_c}
\alias{scale_colour_delphi_d}
\alias{scale_colour_delphi_c}
\title{Delphi palette scales for ggplot2}
\usage{
scale_color_delphi_d(name = "mayfair", direction = 1, ...)

scale_fill_delphi_d(name = "mayfair", direction = 1, ...)

scale_color_delphi_c(name = "mayfair", direction = 1, ...)

scale_fill_delphi_c(name = "mayfair", direction = 1, ...)

scale_colour_delphi_d(name = "mayfair", direction = 1, ...)

scale_colour_delphi_c(name = "mayfair", direction = 1, ...)
}
\arguments{
\item{name}{A single palette name. Use \code{\link[=delphi_palettes]{delphi_palettes()}} to get palette
names. The default is \code{"mayfair"}.}

\item{direction}{The color order. Use \code{1} for the stored order. Use \code{-1} to
reverse the stored order.}

\item{...}{Arguments for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn()}}. Use \code{\link[ggplot2:labs]{ggplot2::labs()}} to set the
legend title.}
}
\value{
A ggplot2 \code{Scale} object.
}
\description{
Uses Delphi palettes in \link[ggplot2:ggplot2]{ggplot2} scales. The discrete scales assign colors
to factor levels. Extra factor levels cause the scales to interpolate colors.
The continuous scales assign colors to numeric values.
\code{\link[=scale_colour_delphi_d]{scale_colour_delphi_d()}} and \code{\link[=scale_colour_delphi_c]{scale_colour_delphi_c()}} are aliases.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)

  ggplot(iris, aes(Species, Sepal.Length, fill = Species)) +
    geom_boxplot() +
    scale_fill_delphi_d("pride")

  ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Petal.Length)) +
    geom_point(size = 2) +
    scale_color_delphi_c("mayfair")
}
}
